[Update]: Recharts tooltip on hover
